Solved Largest Product In Series
Had trouble with the diagonals. Figured out that I can use Numbers to
analyze the data for validation. Then, figured out how to easily show
what the results of traversing the diagonal produced - this output made
it easier to figure out what was wrong with the algorithm and simplified
it.